Move build handler.
